kev::vm_control

Struct VmcsExitCtl

pub struct VmcsExitCtl {
    bits: u32,
}

Table 24-13. Definitions of Primary VM-Exit Controls.

Fields
bits: u32
Implementations

impl VmcsExitCtl

pub const SAVE_DEBUG_CTLS: Self

This control determines whether DR7 and the IA32_DEBUGCTL MSR are saved on VM exit.

The first processors to support the virtual-machine extensions supported only the 1-setting of this control.

pub const HOST_ADDRESS_SPACE_SIZE: Self

On processors that support Intel 64 architecture, this control determines whether a logical processor is in 64-bit mode after the next VM exit. Its value is loaded into CS.L, IA32_EFER.LME, and IA32_EFER.LMA on every VM exit.1 This control must be 0 on processors that do not support Intel 64 architecture.

pub const LOAD_IA32_PERF_GLOBAL_CTRL: Self

This control determines whether the IA32_PERF_GLOBAL_CTRL MSR is loaded on VM exit.

pub const ACK_INTR_ON_EXIT: Self

This control affects VM exits due to external interrupts:

If such a VM exit occurs and this control is 1, the logical processor acknowledges the interrupt controller, acquiring the interruptโs vector. The vector is stored in the VM-exit interruption-information field, which is marked valid.

If such a VM exit occurs and this control is 0, the interrupt is not acknowledged and the VM-exit interruption-information field is marked invalid.

pub const SAVE_IA32_PAT: Self

This control determines whether the IA32_PAT MSR is saved on VM exit.

pub const LOAD_IA32_PAT: Self

This control determines whether the IA32_PAT MSR is loaded on VM exit.

pub const SAVE_IA32_EFER: Self

This control determines whether the IA32_EFER MSR is saved on VM exit.

pub const LOAD_IA32_EFER: Self

This control determines whether the IA32_EFER MSR is loaded on VM exit.

pub const SAVE_VMX_PREEMPTION_TIMER_VALUE: Self

This control determines whether the value of the VMX-preemption timer is saved on VM exit.

pub const CLEAR_IA32_BNDCFGS: Self

This control determines whether the IA32_BNDCFGS MSR is cleared on VM exit.

pub const CONCEAL_VMX_FROM_PT: Self

If this control is 1, Intel Processor Trace does not produce a paging information packet (PIP) on a VM exit or a VMCS packet on an SMM VM exit (see Chapter 32).

pub const CLEAR_IA32_RTIT_CTL: Self

This control determines whether the IA32_RTIT_CTL MSR is cleared on VM exit.

pub const CLEAR_LBR_CTL: Self

This control determines whether the IA32_LBR_CTL MSR is cleared on VM exit.

pub const LOAD_CET_STATE: Self

This control determines whether CET-related MSRs and SPP are loaded on VM exit.

pub const LOAD_PKRS: Self

This control determines whether the IA32_PKRS MSR is loaded on VM exit.

pub const SAVE_IA32_PERF_GLOBAL_CTL: Self

This control determines whether the IA32_PERF_GLOBAL_CTL MSR is saved on VM exit.

pub const ACTIVATE_SECONDARY_CTL: Self

This control determines whether the secondary VM-exit controls are used. If this control is 0, the logical processor operates as if all the secondary VM-exit controls were also 0.

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the โbit-clearโ operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.