Crate kev

Welcome to the KeV project.

Virtualization is an increasingly ubiquitous feature of modern computer systems, and a rapidly evolving part of the system stack. Hardware vendors are adding new features to support more efficient virtualization, OS designs are adapting to perform better in VMs, and VMs are an essential component in cloud computing. Thus, understanding how VMs work is essential to a complete education in computer systems.

In this project, you will skim through the basic components that runs on real virtual machine monitor like KVM. From what you learn, you will build your own type 2 hypervisor and finally extend the hypervisor as an open-ended course project.

In KeV project, we will not bother you from the time-consuming edge case handling and the hidden test cases. The score that you see when run the grading scripts is your final score. We want to keep this project as easy as possible. If you have suggestions on how we can reduce the unnecessary overhead of assignments, cutting them down to the important underlying issues, please let us know.

Projects

The KeV project consists of 5 projects.

KeOS

VMCS and VMexits

Hardware virtualization

Interrupt and I/O virtualization

Final project

Rust

We pick the Rust as a language for project. This is because we believe that after overcome the barriers to learn, memory safety and ownership rule of Rust could significantly reduce the debugging time while implement an operating system.

Getting Started

You can bootstrap your KeV project with following command lines:

$ mkdir keos
$ cd keos
$ curl https://raw.githubusercontent.com/casys-kaist/KeOS/refs/heads/main/scripts/install-kev.sh | sh

PLEASE DO NOT MAKE ANY PUBLIC FORK OF THIS PROJECT. This is strongly denied from the license of the KeV Project. You MUST not redistribute the your works based on the given template.

Enable nested virtualization

See the following docs: https://docs.fedoraproject.org/en-US/quick-docs/using-nested-virtualization-in-kvm/

Additional Resources

Refer to the main KeOS documentation for more details on general workflow of KeV projects, including grading, implementation tips and debugging.

Modules

probe ๐

vcpu

Virtual CPU implementation.

vm

Virtual machine interface.

vm_control

Flags and MSRs for VMX capabilities.

vmcs

Virtual-Machine Control State (VMCS) related apis.

vmexits

Interface to play with vmexits.

Structs

Thread

A Thread for KeV projects.

Enums

VmError

Possible errorkind for Vm.

VmxError

Possible errorkind for Vmx.

Traits

Probe

Traits to probe vcpu internal state.

Functions

start_vmx_on_cpuโ

Enable the VM-eXtension on this cpu.